refactor: refatora formatadores de CPF e data, unificando em columns.formatter
