+++ src/features/acf/modules/AcfDashboardPage/modules/PanelSelector/modules/dashboards/modules/table/modules/diabetes/diabetesAcfDashboardData.adapter.ts
@@ -1,16 +1,18 @@
 import type { impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos } from '@prisma/client';
 import type { ConditionIdentifiedBy, DiabetesAcfItem, PatientAgeRange, PatientStatus } from "./DiabetesAcfItem.model"
+import { isDate, stringToDate } from '@/common/time';
+import { cpf } from "cpf-cnpj-validator";
+
+//essa funcao esta considerando que o cpf é uma string de 11 digitos e não contem o caractere '-' para diferenciar da data de nascimento    
+const isCpf = (possibleCpfString: string): boolean => cpf.isValid(possibleCpfString) && !possibleCpfString.includes("-")
 
 const patientCpfOrBirthdayAdapter = (patientCpfOrBirthdayString : string | null): Date | string | null => {
     if (!patientCpfOrBirthdayString) return null
-    if (patientCpfOrBirthdayString.length === 11 && !patientCpfOrBirthdayString.includes("-")) return patientCpfOrBirthdayString 
-    if (patientCpfOrBirthdayString.length === 10 && patientCpfOrBirthdayString.includes("-")) return new Date(patientCpfOrBirthdayString)
+    if (isCpf(patientCpfOrBirthdayString)) return patientCpfOrBirthdayString 
+    if (isDate(patientCpfOrBirthdayString)) return stringToDate(patientCpfOrBirthdayString)
     return null
 }
-const dateAdapter = (date: Date | null): Date | null => {
-    if (!date) return null
-    return new Date(date.toISOString())
-}
+
 export const diabetesAcfDashboardDataAdapter = (
     data: readonly impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos[]
 ): DiabetesAcfItem[] => {
@@ -19,7 +21,7 @@ export const diabetesAcfDashboardDataAdapter = (
         municipalitySusID: item.municipio_id_sus,
         municipalityState: item.municipio_uf,
         latestExamRequestDate: item.dt_solicitacao_hemoglobina_glicada_mais_recente ? new Date(item.dt_solicitacao_hemoglobina_glicada_mais_recente) : null,
-        mostRecentAppointmentDate: dateAdapter(item.dt_consulta_mais_recente),
+        mostRecentAppointmentDate: item.dt_consulta_mais_recente,
         hemoglobinTestDueDate: item.prazo_proxima_solicitacao_hemoglobina,
         nextAppointmentDueDate: item.prazo_proxima_consulta,
         patientStatus: item.status_usuario as PatientStatus,
@@ -31,6 +33,6 @@ export const diabetesAcfDashboardDataAdapter = (
         careTeamIne: item.equipe_ine_cadastro,
         careTeamName: item.equipe_nome_cadastro,
         visitantCommunityHealthWorker: item.acs_nome_cadastro,
-        mostRecentProductionRecordDate: dateAdapter(item.dt_registro_producao_mais_recente),
+        mostRecentProductionRecordDate: item.dt_registro_producao_mais_recente,
     }))
 }
